+++ hazelcast/src/main/java/com/hazelcast/spi/impl/executionservice/InternalExecutionService.java
@@ -1,0 +1,10 @@
+package com.hazelcast.spi.impl.executionservice;
+
+import com.hazelcast.spi.ExecutionService;
+
+/**
+ * The internal {@link ExecutionService}. Currently this method doesn't contains any additional methods, but
+ * in the future they can be added here.
+ */
+public interface InternalExecutionService extends ExecutionService {
+}
