+++ hazelcast/src/main/java/com/hazelcast/spi/impl/executionservice/impl/ExecutionServiceImpl.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.spi.impl;
+package com.hazelcast.spi.impl.executionservice.impl;
 
 import com.hazelcast.config.ExecutorConfig;
 import com.hazelcast.core.ICompletableFuture;
@@ -22,11 +22,11 @@ import com.hazelcast.instance.HazelcastThreadGroup;
 import com.hazelcast.instance.Node;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.spi.ExecutionService;
-import com.hazelcast.spi.NodeEngine;
 import com.hazelcast.spi.annotation.PrivateApi;
+import com.hazelcast.spi.impl.NodeEngineImpl;
+import com.hazelcast.spi.impl.executionservice.InternalExecutionService;
 import com.hazelcast.util.ConcurrencyUtil;
 import com.hazelcast.util.ConstructorFunction;
-import com.hazelcast.util.ExceptionUtil;
 import com.hazelcast.util.executor.CachedExecutorServiceDelegate;
 import com.hazelcast.util.executor.ExecutorType;
 import com.hazelcast.util.executor.ManagedExecutorService;
@@ -36,15 +36,9 @@ import com.hazelcast.util.executor.SingleExecutorThreadFactory;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.Executor;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Future;
 import java.util.concurrent.LinkedBlockingQueue;
@@ -56,13 +50,10 @@ import java.util.concurrent.SynchronousQueue;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
 
 import static com.hazelcast.util.EmptyStatement.ignore;
 
-public final class ExecutionServiceImpl implements ExecutionService {
+public final class ExecutionServiceImpl implements InternalExecutionService {
 
     private static final int CORE_POOL_SIZE = 3;
     private static final long KEEP_ALIVE_TIME = 60L;
@@ -258,7 +249,7 @@ public final class ExecutionServiceImpl implements ExecutionService {
     }
 
     @PrivateApi
-    void shutdown() {
+    public void shutdown() {
         logger.finest("Stopping executors...");
         for (ExecutorService executorService : executors.values()) {
             executorService.shutdown();
@@ -290,265 +281,5 @@ public final class ExecutionServiceImpl implements ExecutionService {
         CompletableFutureEntry<V> entry = new CompletableFutureEntry<V>(future, nodeEngine);
         completableFutureTask.registerCompletableFutureEntry(entry);
         return entry.completableFuture;
-    }
-
-    private static class CompletableFutureTask implements Runnable {
-        private final List<CompletableFutureEntry> entries = new ArrayList<CompletableFutureEntry>();
-        private final Lock entriesLock = new ReentrantLock();
-
-        private <V> void registerCompletableFutureEntry(CompletableFutureEntry<V> entry) {
-            entriesLock.lock();
-            try {
-                entries.add(entry);
-            } finally {
-                entriesLock.unlock();
-            }
-        }
-
-        @Override
-        public void run() {
-            List<CompletableFutureEntry> removableEntries = removableEntries();
-            removeEntries(removableEntries);
-        }
-
-        private void removeEntries(List<CompletableFutureEntry> removableEntries) {
-            if (removableEntries.isEmpty()) {
-                return;
-            }
-
-            entriesLock.lock();
-            try {
-                entries.removeAll(removableEntries);
-            } finally {
-                entriesLock.unlock();
-            }
-        }
-
-        private List<CompletableFutureEntry> removableEntries() {
-            CompletableFutureEntry[] entries = copyEntries();
-
-            List<CompletableFutureEntry> removableEntries = Collections.EMPTY_LIST;
-            for (CompletableFutureEntry entry : entries) {
-                if (entry.processState()) {
-                    if (removableEntries.isEmpty()) {
-                        removableEntries = new ArrayList<CompletableFutureEntry>(entries.length / 2);
-                    }
-
-                    removableEntries.add(entry);
-                }
-            }
-            return removableEntries;
-        }
-
-        private CompletableFutureEntry[] copyEntries() {
-            if (entries.isEmpty()) {
-                return new CompletableFutureEntry[]{};
-            }
-
-            CompletableFutureEntry[] copy;
-            entriesLock.lock();
-            try {
-                copy = new CompletableFutureEntry[entries.size()];
-                copy = entries.toArray(copy);
-            } finally {
-                entriesLock.unlock();
-            }
-            return copy;
-        }
-    }
-
-    static final class CompletableFutureEntry<V> {
-        private final BasicCompletableFuture<V> completableFuture;
-
-        private CompletableFutureEntry(Future<V> future, NodeEngine nodeEngine) {
-            this.completableFuture = new BasicCompletableFuture<V>(future, nodeEngine);
-        }
-
-        private boolean processState() {
-            if (completableFuture.isDone()) {
-                Object result;
-                try {
-                    result = completableFuture.future.get();
-                } catch (Throwable t) {
-                    result = t;
-                }
-                completableFuture.setResult(result);
-                return true;
-            }
-            return false;
-        }
-    }
-
-    private static final class ScheduledExecutorServiceDelegate implements ScheduledExecutorService {
-
-        private final ScheduledExecutorService scheduledExecutorService;
-        private final ExecutorService executor;
-
-        private ScheduledExecutorServiceDelegate(ScheduledExecutorService scheduledExecutorService, ExecutorService executor) {
-            this.scheduledExecutorService = scheduledExecutorService;
-            this.executor = executor;
-        }
-
-        @Override
-        public void execute(Runnable command) {
-            executor.execute(command);
-        }
-
-        @Override
-        public <T> Future<T> submit(Callable<T> task) {
-            return executor.submit(task);
-        }
-
-        @Override
-        public <T> Future<T> submit(Runnable task, T result) {
-            return executor.submit(task, result);
-        }
-
-        @Override
-        public Future<?> submit(Runnable task) {
-            return executor.submit(task);
-        }
-
-        @Override
-        public ScheduledFuture<?> schedule(Runnable command, long delay, TimeUnit unit) {
-            return scheduledExecutorService.schedule(
-                    new ScheduledTaskRunner(command, executor), delay, unit);
-        }
-
-        @Override
-        public ScheduledFuture<?> scheduleAtFixedRate(Runnable command, long initialDelay, long period, TimeUnit unit) {
-            return scheduledExecutorService.scheduleAtFixedRate(
-                    new ScheduledTaskRunner(command, executor), initialDelay, period, unit);
-        }
-
-        @Override
-        public ScheduledFuture<?> scheduleWithFixedDelay(Runnable command, long initialDelay, long delay, TimeUnit unit) {
-            return scheduledExecutorService.scheduleWithFixedDelay(
-                    new ScheduledTaskRunner(command, executor), initialDelay, delay, unit);
-        }
-
-        @Override
-        public void shutdown() {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public List<Runnable> shutdownNow() {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public boolean isShutdown() {
-            return false;
-        }
-
-        @Override
-        public boolean isTerminated() {
-            return false;
-        }
-
-        @Override
-        public boolean awaitTermination(long timeout, TimeUnit unit) throws InterruptedException {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks)
-                throws InterruptedException {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public <T> List<Future<T>> invokeAll(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
-                throws InterruptedException {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public <T> T invokeAny(Collection<? extends Callable<T>> tasks)
-                throws InterruptedException, ExecutionException {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public <T> T invokeAny(Collection<? extends Callable<T>> tasks, long timeout, TimeUnit unit)
-                throws InterruptedException, ExecutionException, TimeoutException {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public <V> ScheduledFuture<V> schedule(final Callable<V> callable, long delay, TimeUnit unit) {
-            throw new UnsupportedOperationException();
-        }
-    }
-
-    static class BasicCompletableFuture<V> extends AbstractCompletableFuture<V> {
-
-        private final Future<V> future;
-
-        BasicCompletableFuture(Future<V> future, NodeEngine nodeEngine) {
-            super(nodeEngine, nodeEngine.getLogger(BasicCompletableFuture.class));
-            this.future = future;
-        }
-
-        @Override
-        public boolean cancel(boolean mayInterruptIfRunning) {
-            return future.cancel(mayInterruptIfRunning);
-        }
-
-        @Override
-        public boolean isCancelled() {
-            return future.isCancelled();
-        }
-
-        @Override
-        public boolean isDone() {
-            boolean done = future.isDone();
-            if (done && !super.isDone()) {
-                forceSetResult();
-                return true;
-            }
-            return done || super.isDone();
-        }
-
-        @Override
-        public V get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-            V result = future.get(timeout, unit);
-            // If not yet set by CompletableFuture task runner, we can go for it!
-            if (!super.isDone()) {
-                setResult(result);
-            }
-            return result;
-        }
-
-        private void forceSetResult() {
-            Object result;
-            try {
-                result = future.get();
-            } catch (Throwable t) {
-                result = t;
-            }
-            setResult(result);
-        }
-    }
-
-    private static class ScheduledTaskRunner implements Runnable {
-
-        private final Executor executor;
-        private final Runnable runnable;
-
-        public ScheduledTaskRunner(Runnable runnable, Executor executor) {
-            this.executor = executor;
-            this.runnable = runnable;
-        }
-
-        @Override
-        public void run() {
-            try {
-                executor.execute(runnable);
-            } catch (Throwable t) {
-                ExceptionUtil.sneakyThrow(t);
-            }
-        }
     }
 }
